Deprecation Warnings are gone and all tests are greeen except for one mongo mapper spec
The active record side should be okay. But I don't know anything about
mongo mapper. The gemfile is pointed to the github master branch
because it doesn't look like a Rails4 compatible version of the gem has
dropped yet. The failing spec is complaining about a duplicate _id when
rolling back to a specific revision.